+++ package.json
@@ -27,6 +27,8 @@
     "@emotion/styled": "^11.0.0",
     "@nexus/schema": "^0.19.2",
     "@prisma/client": "^2.12.1",
+    "@sanity/client": "^2.1.4",
+    "@sanity/image-url": "^0.140.19",
     "apollo-server-micro": "^2.18.1",
     "date-fns": "^2.16.1",
     "focus-visible": "^5.2.0",
